Fix: app.test.js started to throw warnings due to react-redux update
